Make working save a category and re-use code in UserController
